Chunter: FileBrowser - extract attachments views
